requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
alembic
python-dotenv
boto3
python-magic
argon2-cffi
clamd
Flask-WTF>=1.0
jsonschema>=4.0
